Do not version private API methods
(cherry picked from commit 6018baefd41801eae020ca4c97388e8dd40a0cc0)
