--- src/middlewared/middlewared/api/v25_04_0/netdata.py
@@ -1,25 +1,0 @@
-from pydantic import Field
-
-from middlewared.api.base import BaseModel, NonEmptyString
-
-
-class ChartMetricsDataArgs(BaseModel):
-    before: int = 0
-    after: int = -1
-
-
-class ChartMetricsArgs(BaseModel):
-    chart: NonEmptyString
-    data: ChartMetricsDataArgs = Field(default_factory=lambda: ChartMetricsDataArgs())
-
-
-class ChartMetricsResult(BaseModel):
-    result: dict
-
-
-class ChartDetailsArgs(BaseModel):
-    chart: NonEmptyString
-
-
-class ChartDetailsResult(BaseModel):
-    result: dict
