+++ src/middlewared/middlewared/api/v25_04_0/smb.py
@@ -21,7 +21,6 @@ __all__ = [
     'SmbServiceUnixCharsetChoicesArgs', 'SmbServiceUnixCharsetChoicesResult',
     'SmbServiceBindIPChoicesArgs', 'SmbServiceBindIPChoicesResult',
     'SmbSharePresetsArgs', 'SmbSharePresetsResult',
-    'SmbSharePrecheckArgs', 'SmbSharePrecheckResult',
 ]
 
 EMPTY_STRING = ''
@@ -190,12 +189,3 @@ class SmbSharePresetsArgs(BaseModel):
 
 class SmbSharePresetsResult(BaseModel):
     result: dict[str, dict]
-
-
-@single_argument_args('smb_share_precheck')
-class SmbSharePrecheckArgs(BaseModel):
-    name: NonEmptyString | None = None
-
-
-class SmbSharePrecheckResult(BaseModel):
-    result: Literal[None]
