+++ src/middlewared/middlewared/api/v25_04_0/virt_instance.py
@@ -219,7 +219,6 @@ class ImageChoiceItem(BaseModel):
     archs: list[str]
     variant: str
     instance_types: list[InstanceType]
-    secureboot: bool | None
 
 
 class VirtInstanceImageChoicesResult(BaseModel):
